Add readme
Yes, it’s trying to game the “# of days w/ contribution” system, but
better to start the shell repo and add to it a little at a time.  This
way, at least, someone can let me know early if my coding style is bad
for Swift, so I don’t get in trouble later.